GS-66: Fix Charts on Reports
Changing rendering to one of the charts (Bar Chart, Line Chart, Area Chart,
etc), caused an error to be thrown. This happened because there was an
incompatibility between the pivottable library and recent d3 renderers.

Fixed by reverting dr to v3.5.17 and implementing mandatory version on package
version update script.