GS-71: Fix Issues With Relative Dates Filters
Filtering of some reltive dates was failing because hours of start and end
were not being taken into account. Thus, for fields that were of both date and
time, filters failed, as time taken for both start and end dates was 00:00.
Fixed by including start and end times in relative filter.

On the other hand, specification of the relative dates control required only
dates within the selected range to be shown, however this was not implemented,
merely unchecking dates out of the range. Implemented the required behaviour
by adding .hide() and .show() calls were appropriate.