GS-94: Fix Custom Values not Resolved to Label
Multi select custom values were not being resolved to it's appropriate labels.
This happened because first, the value being passed was the key of the array,
which was always sequential, starting at 0 and incrementing for each value in
the array. This also caused the first value to always be empty, showing as a
starting comma on the report, as none of the option values have 0 as value.

On the other hand, if the field was a string type field, it never got to
resolve its label, as it was first handled as a string with no subsequent
processing.

Fixed by sending the value in the multiple array to be resolved, and handling
optionvalues first when processing values.