GS-130: Implement Customization of Field Labels per Installation
Implemented the possibility to add json files to define a list of labels that
should be changed. These files are to be added per installation, allowing
customization of labels according to each organization's needs.

The files should be put on the resources/custom_labels directory, and they
have to be named after the data class that needs labels to be overridden. The
file should contain a json object with originalLabel => newLabel pairs.

Eg:
{
  'Original Label 1': 'New Label 1',
  'Original Label 2': 'New Label 2'
}